profil/requirements.txt
Django>=3.2,<4.0
djangorestframework==3.15.1
django-cors-headers>=4.3.1
